+++ tests/_support/simple_ns_list.py
@@ -1,11 +1,11 @@
 from invoke import task, Collection
 
 @task
-def z_toplevel():
+def z_toplevel(ctx):
     pass
 
 @task
-def subtask():
+def subtask(ctx):
     pass
 
 ns = Collection(z_toplevel, Collection('a', Collection('b', subtask)))