+++ sites/www/changelog.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* :feature:`114` Removed non-contextualized tasks.
 * :bug:`351` Protect against ``run`` deadlocks involving exceptions in I/O
   threads & nontrivial amounts of unread data in the corresponding subprocess
   pipe(s). This situation should now always result in exceptions instead of